// express-backend/src/db/schema.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<number, bigint | number | string, bigint | number | string>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<string, Date | string, Date | string>;
// ...
export interface HostapplicationsHostapplication {
  address: string;
  agreement_document: string | null;
  agreement_version: string;
  ai_rationale: string;
  ai_risk_score: number | null;
  applicant_id: Int8;
  compliance_reviewed_at: Timestamp | null;
  compliance_reviewed_by_id: Int8 | null;
  created_at: Timestamp;
  decline_reason: string;
  declined_stage: string;
  full_name: string;
  headshot: string;
  id: Generated<Int8>;
  id_document: string;
  momo_network: string;
  momo_number: string;
  next_of_kin_name: string;
  next_of_kin_phone: string;
  next_of_kin_relationship: string;
  ps_reviewed_at: Timestamp | null;
  ps_reviewed_by_id: Int8 | null;
  status: string;
  supervisor_reviewed_at: Timestamp | null;
  supervisor_reviewed_by_id: Int8 | null;
  tax_clearance_receipt: string | null;
  updated_at: Timestamp;
}